requirements.txt
torch>=2.0.0
torchvision>=0.15.0
open_clip_torch>=2.31.0
scipy>=1.10.0
numpy>=1.24.0
Pillow>=9.0.0
kagglehub>=1.0.0
